requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.25.0
accelerate>=0.16.0
psycopg2-binary>=2.9.0
Pillow>=9.0.0
numpy>=1.21.0
tqdm>=4.64.0
pandas>=1.5.0
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
webdriver-manager>=4.0.0
lxml>=4.9.0
fake-useragent>=1.1.0
